Some clang-tidy fixes

+++ src/Cool/Exporter/VideoExportProcess.cpp
@@ -32,7 +32,7 @@ void VideoExportProcess::imgui()
 {
     int frame_count = _nb_frames_which_finished_exporting.load();
     ImGui::Text("%s", ("Exported " + String::to_string(frame_count, _max_nb_digits_of_frame_count) + " / " + std::to_string(_total_nb_of_frames_in_sequence) + " frames").c_str());
-    ImGuiExtras::time_formated_hms((_total_nb_of_frames_in_sequence - frame_count) * _frame_time_average / _thread_pool.size());
+    ImGuiExtras::time_formated_hms(static_cast<float>(_total_nb_of_frames_in_sequence - frame_count) * _frame_time_average / static_cast<float>(_thread_pool.size()));
     ImGui::SameLine();
     ImGui::Text("remaining");
     if (ImGui::Button("Stop exporting")) {
